fix(mission): suppress Starting mission pill flash on workspace nav
The cyan "Starting mission…" pill flashed on every mission click
because the load gate at `MissionWorkspace.tsx` blocks the tab strip
until `Promise.all(mission.get, session.list, mission.eventsReplay)`
resolves. The gate is load-bearing — the tabs, feed, and rail below
all assume `mission` is non-null — so we keep it but delay the
*visible* pill: only render after the gate has been blocking for
150ms. Fast IPC (common path) never shows the pill; slow IPC (cold
start, contended SQLite) still gets user feedback.

Same fix shape as #179 for direct-chat nav. Lifted the
`useDelayedFlag` hook out of `RunnerChat.tsx` into
`src/lib/useDelayedFlag.ts` so both pages share one implementation
instead of drifting copies. Reset key is the mission id so
mission-to-mission nav starts a fresh delay window — A→B→A doesn't
reuse a stale fire.
